backend/requirements.txt

# get all lib by freeze 
asgiref==3.8.1
cffi==1.16.0
crispy-bootstrap4==2024.1
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
easy-thumbnails==2.8.5
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1